+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## head:
+* add proxy support to java deployment record LWRP
 
 ## 0.3.1(06/07/2013):
 * internal refactor of proxy handling for server monitor agent